Start advent430 branch for correctness testing.
The purpose of this branch is to create a version of the game from
before the bug fixes, refactoring, and logic changes.  We want this so
we can run it against our 100% coverage test suite and see all changes
in behavior.

This branch is forked from the point where the prompt and the oldstyle
option were added.  At this point there had been only two logic
changes:

1. Do initialization of the LCG with gettimeofday(). Note that
this change will not affectt regression testing, since the
initialization done in this way will nbe overridden in the
logs by seed commands.

2. Refactor the input routines to a normal Unixy organization.
This is required for the -l option to work.

This commit just builds the binary at advent430 where it
won't collide with the production version.